Add CFF Type 2 CharString operator 'flex1'
Issue #85 reported malformed characters displayed from two OTF font files.
Investigation found that the CFF CharString instructions in these fonts
employed the rather unusual operator 'flex1', which was not implemented
by opentype.js.  About 8 out of 300+ characters used the operator.

Upon encountering the 'flex1' operator the code would simply skip over
the operator.  This resulted in the accumulated operand values being
left on the stack, which disrupted execution of following operators, and
left the affected characters obviously distorted.  Clearing of the stack
when an unsupported operator is found was added to fix that.

This still left the characters distorted because one of the drawing
instructions was missing.  (In this case the missing part was vertical
so the resulting curves were not tall enough).

Rereading the documentation induced dread but finally resulted in working
code to implement this operator, though not completely.  When fully
implemented code is supposed to decide whether the Bezier curves will
actually be so shallow that they should just be replaced with a straight
line.  This fix instead always outputs the two Bezier curves.

Tested against the two fonts referenced in issue #85.  (Thank you @georg-paul)
The glyphs 'D' (GID#18) and 'P' (GID#79) were example problems that were
corrected when using font LeagueGothic-Regular.otf  Glyph 't' (GID#294) was
the example problem in font Neris-Light.otf and it was also corrected with
these fixes.

Note that JSHint pointed out the lack of a semicolon from a prior fix #86
and this was added to this fix.