Add support for CID-keyed font in OTF files
This particular patch is incomplete and needs more work.  It is submitted
only for evaluation and soliciting comments.

Also the *major* incomplete item is glyph naming for CID-keyed fonts;
I've forgotten where I read about what convention is used!

opentype.js was not looking for the signal that a CFF table represented
a CID-keyed font, and thus was not capturing the information from the
accompanying FDArray and FDSelect subtables.

In CID-keyed fonts some per-glyph information is found only in the FDArray
subtables, not as usual in the single font-wide Private DICT.  And the
directory of which FDArray[] to use for each glyph is found only in the
one FDSelect lookup list.

The FDArray data layout almost exactly parallels that for the high-level
Top DICT INDEX.  Adding support for FDArray was best done by pulling
Top DICT-related code from parseCFFTable() and creating a new routine to
process an INDEX into a list of 'TopDicts' and their linked Private DICTs.

Various notes:

  - Added a missing check (in parseCFFTable()) against a multi-font
    'FontSet', which is not supported by opentype.js;
        (is this feature used by modern font/code?)

  - as before, only the first Top DICT from Top DICT INDEX is used;

  - I preserved the copying 'defaultWidthX' and 'nominalWidthX' into
    'font' from the TopDict; I think this should be deleted as now
    there could be a different value for each glyph;

  - had to move code for calcCFFSubroutineBias() upwards to satisfy JSHint;

  - now that multiple tables might be created using the default values
    in TOP_DICT_META and PRIVATE_DICT_META, I wonder about the copying
    of the default values that are arrays.  As I read it, multiple
    tables would _share_ the same array values copied from the defaults.
    Perhaps it would be better in interpretDict() to 'clone' the arrays
    with something like "  [].concat(value)  " ?

  - the conditional code at beginning of parseCFFCharstring() ought to
    be shortened to just selecting which TopDict/FDArray[] to then copy
    values from;

  - the multiple " width + nominalWidthX " should just be done once at
    end of parsing the CharString;