Write out CFF fontMatrix.
The default CFF fontMatrix is defined for a font with a UPM of 1000.
Other UPMs need to set a different fontMatrix, but this was never
propagated. We now set the values of the fontMatrix based on the font’s
unitsPerEm. Values other than the default will be encoded in the CFF
top dict.

This fixes #115.