Parse ‘ltag’ table, and pipe it through to name.parse()
The ‘ltag’ table contains IETF BCP-47 language tags, so fonts can
support arbitrary languages. It is basically Apple’s version of what
Microsoft did in “format 1” naming tables of OpenType.

In a later change, I will modify the ‘name’ parser so it actually uses
these language tags (if people are fine with this plan). It should be
straightforward to support both the Apple and the Microsoft way.