Support multilingual names
This change affects the public API to OpenType.js: Instead of
`familyName`, the English font family name is now stored as
`fontFamily.en`, the Spanish name as `fontFamily.es`, and so on.
Likewise for all other names stored in a font.

OpenType names are rather complex to interpret. The API hides this
complexity; language codes are exposed as IETF BCP 47 language tags,
which most web programmers already know from HTML and XML.

Resolves https://github.com/nodebox/opentype.js/issues/144.