Ignore unvisualized tables in font inspector
Before this change, font-inspector.html would sometimes try to
populate an inexistent DOM element, resulting in a JavaScript
exception.  An example where this happened was fonts with an ‘ltag’
table; but language tags will best visualized within the ‘name’ panel,
they do not need their own panel.  Another example for a purely
supportive table (that will not need its own panel for visualization)
is ‘avar’, which is probably best visualized as part of the ‘fvar’
panel.