Encode variation adjustment deltas
This is a port of Python code which I’d written a while ago for
fonttools. It encodes a structure that is part of the `gvar` and
`cvar` tables.  Actually I had thought I’d sent this out for review
back when I’d written the code, but it looks like it has been silently
sitting in my git client.

Sadly, I won’t have time to contribute a full `gvar` implementation to
OpenType.js, but perhaps you’ll find this small piece still useful.