Better errors for unsupported cmap sub-tables.
Before, if we couldn't find a sub-table with a supported platform id /
encoding id, we returned null in cmap.parse, which confused the code
later on, resulting in an "Uncaught TypeError: Cannot read property
'glyphIndexMap' of null" error.

Now, we immediately throw when we can't find a valid sub-table.

In addition, if the format is unsupported, we display which format we
got.