Fix issue with argument default value. Add two test fonts.
This fixes the issue with specifying a default value in parseVersion
function argument. This also adds two test fonts. TestTracking3.ttf may
be used for testing fonts with GPOS table version 1.1. This variable
font allows the user to adjust the character tracking using variations
in conjunction with GPOS variations. The TestItalicSwitch.ttf font is
used to test fonts with GSUB table version 1.1. This variable font
allows the user to switch between upright and italic styles using a
contextural substitution based on the italic variations axis.