fix: update isNode implemenation
isNode previously assumed that an environment without a window was 'node'.   Unfortunately is also suggests a 'worker' environment.

In the context of the code, a node environment has access to the file-system, but a worker is more like the browser environment.

Updated isNode to check the presence of values in the node environment, instead of levaging the lack of a window.

This should fix issues with fonts attempting to read via readFileSync instead of XHR/fetch