cff: Implement private DICT reading and writing for CFF and CFF2
[why]
The hinting of CFF (1 and 2) Open Type fonts is usually
entirely in their private DICT data. To preserve the
hinting - which is needed to make the font rendering look
good and as expected - the private DICT data needs to be
read and written.

[how]
Parts of the needed code seem to be added already in
preparation for CFF2. I guess there was a misunderstanding
or a mixture of versions, but the private DICT did not
change between CFF1 and CFF2, and we need to always use
the PRIVATE_DICT_META_CFF2, the link given in its
definition is the same as given with link [1] for CFF1.
See also [2]. So firstly we always refer to 'version 2',
as the previous code did also but only in one case.

The OtherBlues and FamilyOtherBlues operators must occur
after the BlueValues and FamilyBlues operators, respectively.
This is implicitely guaranteed by the way the private DICT is
set up and finally Object.keys() works.

For writing the delta values the encoding function is missing
and so that is added. encode.delta() just calls encode.number()
repeatedly, figuratively speaking.

Last but not least the correct private DICT length has to be set.

[1] https://adobe-type-tools.github.io/font-tech-notes/pdfs/5176.CFF.pdf
[2] https://learn.microsoft.com/en-us/typography/opentype/otspec180/cff2#appendixD

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>