cff: Humanify user facing way "delta"s are handled
[why]
At the moment the delta values are not de- or encoded. The delta format
is a packed format to have the smallest possible values in the array of
numbers (i.e. relative to the previous number).
But usually users do not face this but are shown absolute values; in
all other font specific applications.

For example
BlueValues [ 500, 550 ]      // User sees the BlueZone is from 500 to 550
Encoded as [ 500, 50 ]

opentype.js at the moment does not translate these deltas in any way and
users must know this and use the inconvenient 'packed' encoding format.

[how]
Convert the read relative delta values to absolute coordinates in the
blueValues (and other) properties, that users can interact with.

On font encoding time the absolute coordinates are converted back to
relative ones and encoded in the font file.

[note]
https://adobe-type-tools.github.io/font-tech-notes/pdfs/5176.CFF.pdf

    The second and subsequent numbers in a delta are encoded as the
    difference between successive values. For example, an array a0,
    a1, ..., an would be encoded as: a0 (a1–a0) (a2–a1) ..., (an–a(n–1))

This is done because small numbers can be encoded with fewer bytes and
the total aim is to reduce the size.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>