cff: Fix not-handling subrs
[why]
The code does not handle subrs, but it includes a previously existing
subrs operator in the private DICT which throws all the offsets off.

[how]
Either include the subrs subtable and keep the subrs operator, or drop
both. This commits just drops the operator (for simplicity).

Proper subrs handling can be added later on.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>