cff: test: Roundtrip check of private DICT
[why]
Writing of the private DICT has been added together with the translation
of delta values from transport encoding to readable and back. But there
is no automated test to see if something breaks.

[how]
Take one of the fonts in test/fonts/ and run in through ttx to see its
private DICT data analyzed by someone else.

Add a test that also reads that fond and checks some of the private DICT
values.

Write the font into a buffer and read it back in; repeat the check on
the read back font. This also checks if writing is correct.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>