cff: Revert to use different private DICTs for CFF and CFF2
[why]
While CFF has no `vsindex` operator in the private DICT it is a
defaulted operator in CFF2. This can not be really disentangled with the
current code and the opertor ends up erroreously written to CFF fonts
(where it is usually ignored).

[how]
It took a bit time to understand the CFF version code (that I believe is
not finished), but reinstate that code with different tables to be used.

Sorry I messed with the code in the first place; I had overlooked the
`vsindex` operator and how a defaulted operator in the tables turn out
in the written files.

I just reverted the version 'detection' code as it was, not checking
anything. Note that there still is one hard-coded `2` somewhere, but
that was already there when I started.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>