cff: Fix version number
[why]
The version number written out to a buffer is not the number stored in
font.tables.cff.topDict.version. Instead some other human readable
string is used even if the cff.topDict.version has been set correctly.

The CFF version number needs to be a string representing a number and
nothing else; especially it should not contain 'Version'.

Documentation [1] says this:
    version 0 SID
    This can be derived by copying the OpenType 'head' fsRevision
    field, and formatting it as a Fixed with three decimal places of
    precision.

The documentation is also valid for CFF1 in this regard.

[how]
Use the version string in the CFF table, if there is one.
If not fall back to the (imho wrong) previous defaulting behavior.

[1] https://learn.microsoft.com/en-us/typography/opentype/otspec180/cff2#sectionD3

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>